+++ gen/src/check.rs
@@ -40,9 +40,11 @@ pub(crate) fn run(args: CheckArgs) -> Result<bool> {
             if parser.warnings().is_empty() && parser.other_warnings().is_empty() {
                 eprintln!("No issues found.");
             } else {
+                eprintln!();
                 eprintln!("Some warning(s) found.");
             }
         } else {
+            eprintln!();
             eprintln!("Some error(s) found.");
             res = false;
         }