+++ gen/src/rust.rs
@@ -52,9 +52,11 @@ pub(crate) fn run(args: RustArgs) -> Result<bool> {
 
     if parser.errors().is_empty() {
         if !parser.warnings().is_empty() || !parser.other_warnings().is_empty() {
+            eprintln!();
             eprintln!("Some warning(s) found.");
         }
     } else {
+        eprintln!();
         eprintln!("Some error(s) found.");
         return Ok(false);
     }